Moves build-utils.js to build/utils.js
